refactor VideoCard.js로 rename
