+++ src/js/common/VideoCard/VideoCard.js
